Moved internalClient to its own file.
Eventually, there will be a `v3/client.go`, `v3/client_internal.go` and
`v3/client_sortkey.go`, where all the shared logic is housed in
`v3/client_internal.go`. This is the first step of that: separating out
`internalClient` from `Client`.
